[BluetoothAudioSink] Rename local interface files
